refactor: move tokenizer classes inside tokenizer package
