+++ src/main/java/com/computer/nand2tetris/compiler/JackToken.java
@@ -3,13 +3,13 @@ package com.computer.nand2tetris.compiler;
 import com.google.auto.value.AutoValue;
 
 @AutoValue
-abstract class JackToken {
+public abstract class JackToken {
 
-  enum TokenType {
+  public enum TokenType {
     SYMBOL, INTEGER_CONSTANT, STRING_CONSTANT, KEYWORD, IDENTIFIER,
   }
 
-  static JackToken create(TokenType tokenType, String tokenText) {
+  public static JackToken create(TokenType tokenType, String tokenText) {
     return new AutoValue_JackToken(tokenType, tokenText);
   }
